requirements.txt
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
terminaltables==3.1.0
